// enum/error-type.ts
export enum ErrorMessage {
  FirstName = "First name is required!",
  LastName = "Lase name is required!",
  Name = "Name is required!",
  EmailValidation = "Please enter a valid email.",
  UsernameValidation = "Username must be at least 4 characters long.",
  PswValidation = "Password must contain at least one uppercase letter, one lowercase letter, one number, and one special character and must be at least 8 characters long.",
  RoleDoesNotExist = "The role does not exist.",
  Db = "Database error",
  PswNotMatch = "Passwords do not match.",
  EmailExist = "E-Mail already exists!",
  UsernameExist = "Username already exists!",
  EmailNotExist = "E-Mail does not  exists!",
}